+++ scripts/serve/serve.js
@@ -14,7 +14,7 @@ const DEV_SERVER_HOST = Object.prototype.hasOwnProperty.call(process.env, 'DEV_S
   : '0.0.0.0';
 const DEV_SERVER_PORT = Object.prototype.hasOwnProperty.call(process.env, 'DEV_SERVER_PORT')
   ? parseInt(process.env.DEV_SERVER_PORT, 10)
-  : 8080;
+  : 8081;
 
 async function main () {
   const webpackDevServer = WebpackDevServer.factory();